refactor(css): remove the now-dead CssIcssExportDependency
The per-export dependency is no longer created since exports are consolidated
into CssExportsDependency. Move the still-used pieces (getLocalIdent,
computeInterpolatedIdentifier, and the EXPORT_MODE/EXPORT_TYPE enums) into
CssExportsDependency, repoint the ExportMode/ExportType typedefs, drop the
serialization and dependency-template registrations, and delete the file.
