+++ lib/css/CssModulesPlugin.js
@@ -26,7 +26,6 @@ const NormalModule = require("../NormalModule");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
 const CssExportsDependency = require("../dependencies/CssExportsDependency");
-const CssIcssExportDependency = require("../dependencies/CssIcssExportDependency");
 const CssIcssImportDependency = require("../dependencies/CssIcssImportDependency");
 const CssIcssSymbolDependency = require("../dependencies/CssIcssSymbolDependency");
 const CssImportDependency = require("../dependencies/CssImportDependency");
@@ -251,10 +250,6 @@ class CssModulesPlugin {
 				compilation.dependencyTemplates.set(
 					CssIcssImportDependency,
 					new CssIcssImportDependency.Template()
-				);
-				compilation.dependencyTemplates.set(
-					CssIcssExportDependency,
-					new CssIcssExportDependency.Template()
 				);
 				compilation.dependencyTemplates.set(
 					CssExportsDependency,
