+++ lib/css/CssParser.js
@@ -11,7 +11,6 @@ const { CSS_MODULE_TYPE_AUTO } = require("../ModuleTypeConstants");
 const Parser = require("../Parser");
 const ConstDependency = require("../dependencies/ConstDependency");
 const CssExportsDependency = require("../dependencies/CssExportsDependency");
-const CssIcssExportDependency = require("../dependencies/CssIcssExportDependency");
 const CssIcssImportDependency = require("../dependencies/CssIcssImportDependency");
 const CssIcssSymbolDependency = require("../dependencies/CssIcssSymbolDependency");
 const CssImportDependency = require("../dependencies/CssImportDependency");
@@ -616,11 +615,11 @@ class CssParser extends Parser {
 		 * @param {number} el end line
 		 * @param {number} ec end column
 		 * @param {string} name export name
-		 * @param {import("../dependencies/CssIcssExportDependency").Value} value value or [localName, importName, request?]
+		 * @param {import("../dependencies/CssExportsDependency").Value} value value or [localName, importName, request?]
 		 * @param {Range=} range source range to replace, when present
 		 * @param {boolean=} interpolate whether the value needs interpolation
-		 * @param {import("../dependencies/CssIcssExportDependency").ExportMode=} exportMode export mode
-		 * @param {import("../dependencies/CssIcssExportDependency").ExportType=} exportType export type
+		 * @param {import("../dependencies/CssExportsDependency").ExportMode=} exportMode export mode
+		 * @param {import("../dependencies/CssExportsDependency").ExportType=} exportType export type
 		 * @returns {void}
 		 */
 		const addCssExport = (
@@ -632,8 +631,8 @@ class CssParser extends Parser {
 			value,
 			range,
 			interpolate = false,
-			exportMode = CssIcssExportDependency.EXPORT_MODE.REPLACE,
-			exportType = CssIcssExportDependency.EXPORT_TYPE.NORMAL
+			exportMode = CssExportsDependency.EXPORT_MODE.REPLACE,
+			exportType = CssExportsDependency.EXPORT_TYPE.NORMAL
 		) => {
 			cssExportEntries.push({
 				name,
@@ -1223,10 +1222,10 @@ class CssParser extends Parser {
 						getReexport(identifier),
 						[cv.start, cv.end],
 						true,
-						CssIcssExportDependency.EXPORT_MODE.ONCE,
+						CssExportsDependency.EXPORT_MODE.ONCE,
 						isDashedIdent
-							? CssIcssExportDependency.EXPORT_TYPE.CUSTOM_VARIABLE
-							: CssIcssExportDependency.EXPORT_TYPE.NORMAL
+							? CssExportsDependency.EXPORT_TYPE.CUSTOM_VARIABLE
+							: CssExportsDependency.EXPORT_TYPE.NORMAL
 					);
 				}
 			}
@@ -1266,7 +1265,7 @@ class CssParser extends Parser {
 							value,
 							[cv.start, cv.end],
 							true,
-							CssIcssExportDependency.EXPORT_MODE.ONCE
+							CssExportsDependency.EXPORT_MODE.ONCE
 						);
 						found = true;
 						pure.markLocal();
@@ -1297,8 +1296,8 @@ class CssParser extends Parser {
 							getReexport(identifier),
 							[cv.start, cv.end],
 							true,
-							CssIcssExportDependency.EXPORT_MODE.ONCE,
-							CssIcssExportDependency.EXPORT_TYPE.NORMAL
+							CssExportsDependency.EXPORT_MODE.ONCE,
+							CssExportsDependency.EXPORT_TYPE.NORMAL
 						);
 						found = true;
 						pure.markLocal();
@@ -1348,8 +1347,8 @@ class CssParser extends Parser {
 				getReexport(identifier, undefined, true),
 				[identStart, identEnd],
 				true,
-				CssIcssExportDependency.EXPORT_MODE.ONCE,
-				CssIcssExportDependency.EXPORT_TYPE.CUSTOM_VARIABLE
+				CssExportsDependency.EXPORT_MODE.ONCE,
+				CssExportsDependency.EXPORT_TYPE.CUSTOM_VARIABLE
 			);
 		};
 
@@ -1394,8 +1393,8 @@ class CssParser extends Parser {
 				getReexport(identifier, localName, true),
 				[identStart, sourceEnd - 1],
 				true,
-				CssIcssExportDependency.EXPORT_MODE.ONCE,
-				CssIcssExportDependency.EXPORT_TYPE.CUSTOM_VARIABLE
+				CssExportsDependency.EXPORT_MODE.ONCE,
+				CssExportsDependency.EXPORT_TYPE.CUSTOM_VARIABLE
 			);
 
 			module.addPresentationalDependency(
@@ -1709,7 +1708,7 @@ class CssParser extends Parser {
 						getReexport(idName),
 						[idValueStart, v.end],
 						true,
-						CssIcssExportDependency.EXPORT_MODE.ONCE
+						CssExportsDependency.EXPORT_MODE.ONCE
 					);
 					pure.markLocal();
 					continue;
@@ -1794,7 +1793,7 @@ class CssParser extends Parser {
 						getReexport(className),
 						[classNameStart, classNameEnd],
 						true,
-						CssIcssExportDependency.EXPORT_MODE.NONE
+						CssExportsDependency.EXPORT_MODE.NONE
 					);
 					continue;
 				}
@@ -1832,7 +1831,7 @@ class CssParser extends Parser {
 							getReexport(name),
 							[next.start, next.end],
 							true,
-							CssIcssExportDependency.EXPORT_MODE.ONCE
+							CssExportsDependency.EXPORT_MODE.ONCE
 						);
 						currentRule.hasLocalAnchor = true;
 						currentRule.localIdentifiers.push(name);
@@ -2403,7 +2402,7 @@ class CssParser extends Parser {
 												getReexport(localName),
 												undefined,
 												false,
-												CssIcssExportDependency.EXPORT_MODE.REPLACE
+												CssExportsDependency.EXPORT_MODE.REPLACE
 											);
 										}
 									}
@@ -2925,8 +2924,8 @@ class CssParser extends Parser {
 										getReexport(identifier),
 										[start, end],
 										true,
-										CssIcssExportDependency.EXPORT_MODE.SELF_REFERENCE,
-										CssIcssExportDependency.EXPORT_TYPE.COMPOSES
+										CssExportsDependency.EXPORT_MODE.SELF_REFERENCE,
+										CssExportsDependency.EXPORT_TYPE.COMPOSES
 									);
 									continue;
 								}
@@ -2955,8 +2954,8 @@ class CssParser extends Parser {
 									getReexport(identifier, localName),
 									[start, end],
 									true,
-									CssIcssExportDependency.EXPORT_MODE.APPEND,
-									CssIcssExportDependency.EXPORT_TYPE.COMPOSES
+									CssExportsDependency.EXPORT_MODE.APPEND,
+									CssExportsDependency.EXPORT_TYPE.COMPOSES
 								);
 							}
 						} else if (fromSource && fromSource.kind === "global") {
@@ -2973,8 +2972,8 @@ class CssParser extends Parser {
 									getReexport(identifier),
 									[start, end],
 									false,
-									CssIcssExportDependency.EXPORT_MODE.APPEND,
-									CssIcssExportDependency.EXPORT_TYPE.COMPOSES
+									CssExportsDependency.EXPORT_MODE.APPEND,
+									CssExportsDependency.EXPORT_TYPE.COMPOSES
 								);
 							}
 						} else {
@@ -2992,9 +2991,9 @@ class CssParser extends Parser {
 									[start, end],
 									!isGlobal,
 									isGlobal
-										? CssIcssExportDependency.EXPORT_MODE.APPEND
-										: CssIcssExportDependency.EXPORT_MODE.SELF_REFERENCE,
-									CssIcssExportDependency.EXPORT_TYPE.COMPOSES
+										? CssExportsDependency.EXPORT_MODE.APPEND
+										: CssExportsDependency.EXPORT_MODE.SELF_REFERENCE,
+									CssExportsDependency.EXPORT_TYPE.COMPOSES
 								);
 							}
 						}
@@ -3136,10 +3135,10 @@ class CssParser extends Parser {
 							getReexport(name),
 							[start, end],
 							true,
-							CssIcssExportDependency.EXPORT_MODE.ONCE,
+							CssExportsDependency.EXPORT_MODE.ONCE,
 							isGridProperty
-								? CssIcssExportDependency.EXPORT_TYPE.GRID_CUSTOM_IDENTIFIER
-								: CssIcssExportDependency.EXPORT_TYPE.NORMAL
+								? CssExportsDependency.EXPORT_TYPE.GRID_CUSTOM_IDENTIFIER
+								: CssExportsDependency.EXPORT_TYPE.NORMAL
 						);
 					}
 				}
