+++ lib/dependencies/CssExportsDependency.js
@@ -6,11 +6,15 @@
 "use strict";
 
 const { JAVASCRIPT_TYPE } = require("../ModuleSourceTypeConstants");
+const { interpolate } = require("../TemplatedPathPlugin");
 const WebpackError = require("../errors/WebpackError");
 const { cssExportConvention } = require("../util/conventions");
+const createHash = require("../util/createHash");
+const { makePathsRelative } = require("../util/identifier");
 const makeSerializable = require("../util/makeSerializable");
 const memoize = require("../util/memoize");
-const CssIcssExportDependency = require("./CssIcssExportDependency");
+const nonNumericOnlyHash = require("../util/nonNumericOnlyHash");
+const { updateHashFromSource } = require("../util/source");
 const CssIcssImportDependency = require("./CssIcssImportDependency");
 const NullDependency = require("./NullDependency");
 
@@ -18,6 +22,8 @@ const getCssTokens = memoize(() => require("../css/walkCssTokens"));
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
 /** @typedef {import("../../declarations/WebpackOptions").CssGeneratorExportsConvention} CssGeneratorExportsConvention */
+/** @typedef {import("../../declarations/WebpackOptions").CssGeneratorLocalIdentName} CssGeneratorLocalIdentName */
+/** @typedef {import("../../declarations/WebpackOptions").HashFunction} HashFunction */
 /** @typedef {import("../css/CssGenerator")} CssGenerator */
 /** @typedef {import("../css/CssModule")} CssModule */
 /** @typedef {import("../ChunkGraph")} ChunkGraph */
@@ -30,24 +36,187 @@ const getCssTokens = memoize(() => require("../css/walkCssTokens"));
 /** @typedef {import("../RuntimeTemplate")} RuntimeTemplate */
 /** @typedef {import("../css/CssParser").Range} Range */
 /** @typedef {import("../Dependency").DependencyLocation} DependencyLocation */
-/** @typedef {import("./CssIcssExportDependency").ExportMode} ExportMode */
-/** @typedef {import("./CssIcssExportDependency").ExportType} ExportType */
-/** @typedef {import("./CssIcssExportDependency").Value} Value */
 /** @typedef {import("../serialization/ObjectMiddleware").ObjectDeserializerContext} ObjectDeserializerContext */
 /** @typedef {import("../serialization/ObjectMiddleware").ObjectSerializerContext} ObjectSerializerContext */
 /** @typedef {import("../util/Hash")} Hash */
 /** @typedef {import("../util/runtime").RuntimeSpec} RuntimeSpec */
 /** @typedef {import("../errors/WebpackError")} WebpackErrorType */
 
-const EXPORT_MODE = CssIcssExportDependency.EXPORT_MODE;
-const EXPORT_TYPE = CssIcssExportDependency.EXPORT_TYPE;
+/** @typedef {string | [string, string] | [string, string, string]} Value */
+// 0 - none, 1 - replace, 2 - append, 3 - once, 4 - self-reference
+/** @typedef {0 | 1 | 2 | 3 | 4} ExportMode */
+// 0 - normal, 1 - custom css variable, 2 - grid custom ident, 3 - composes
+/** @typedef {0 | 1 | 2 | 3} ExportType */
+
+/** @type {Record<"NONE" | "REPLACE" | "APPEND" | "ONCE" | "SELF_REFERENCE", ExportMode>} */
+const EXPORT_MODE = {
+	NONE: 0,
+	REPLACE: 1,
+	APPEND: 2,
+	ONCE: 3,
+	SELF_REFERENCE: 4
+};
+
+/** @type {Record<"NORMAL" | "CUSTOM_VARIABLE" | "GRID_CUSTOM_IDENTIFIER" | "COMPOSES", ExportType>} */
+const EXPORT_TYPE = {
+	NORMAL: 0,
+	CUSTOM_VARIABLE: 1,
+	GRID_CUSTOM_IDENTIFIER: 2,
+	COMPOSES: 3
+};
 
 /**
- * One `:export { name: value }` style entry. Same shape as the legacy
- * per-export `CssIcssExportDependency` fields, but a plain object so a module
- * holds a single `Dependency` plus a flat array instead of thousands of
- * `Dependency` instances. `_conventionNames`/`_valueConventionNames` memoize
- * the convention expansion the way the legacy class did per instance.
+ * Returns local ident.
+ * @param {string} local css local
+ * @param {CssModule} module module
+ * @param {ChunkGraph} chunkGraph chunk graph
+ * @param {RuntimeTemplate} runtimeTemplate runtime template
+ * @returns {string} local ident
+ */
+const getLocalIdent = (local, module, chunkGraph, runtimeTemplate) => {
+	const generator = /** @type {CssGenerator} */ (module.generator);
+	const localIdentName =
+		/** @type {CssGeneratorLocalIdentName} */
+		(generator.options.localIdentName);
+	const relativeResourcePath = makePathsRelative(
+		/** @type {string} */
+		(runtimeTemplate.compilation.compiler.context),
+		/** @type {string} */
+		(module.getResource()),
+		runtimeTemplate.compilation.compiler.root
+	);
+	const { uniqueName } = runtimeTemplate.outputOptions;
+
+	let localIdentHash = "";
+
+	if (generator._localIdentNeedsHash) {
+		const hashSalt = generator.options.localIdentHashSalt;
+		const hashDigest =
+			/** @type {string} */
+			(generator.options.localIdentHashDigest);
+		const hashDigestLength = generator.options.localIdentHashDigestLength;
+		const hashFunction =
+			/** @type {HashFunction} */
+			(generator.options.localIdentHashFunction);
+
+		const hash = createHash(hashFunction);
+
+		if (hashSalt) {
+			hash.update(hashSalt);
+		}
+
+		if (uniqueName) {
+			hash.update(uniqueName);
+		}
+
+		hash.update(relativeResourcePath);
+		hash.update(local);
+
+		localIdentHash = hash.digest(hashDigest).slice(0, hashDigestLength);
+	}
+
+	let contentHash = "";
+
+	if (generator._localIdentNeedsContentHash) {
+		const hash = createHash(runtimeTemplate.outputOptions.hashFunction);
+		const source = module.originalSource();
+
+		if (source) {
+			updateHashFromSource(hash, source);
+		}
+
+		if (module.error) {
+			hash.update(module.error.toString());
+		}
+
+		const fullContentHash = hash.digest(
+			runtimeTemplate.outputOptions.hashDigest
+		);
+
+		contentHash = nonNumericOnlyHash(
+			fullContentHash,
+			runtimeTemplate.outputOptions.hashDigestLength
+		);
+	}
+
+	let localIdent = interpolate(localIdentName, {
+		prepareId: (id) => {
+			if (typeof id !== "string") return id;
+
+			return (
+				id
+					.replace(/^([.-]|[^a-z0-9_-])+/i, "")
+					// We keep the `@` symbol because it can be used in the package name (e.g. `@company/package`), and if we replace it with `_`, a class conflict may occur.
+					// For example - `@import "@foo/package/style.module.css"` and `@import "foo/package/style.module.css"` (`foo` is a package, `package` is just a directory) will create a class conflict.
+					.replace(/[^a-z0-9@_-]+/gi, "_")
+			);
+		},
+		filename: relativeResourcePath,
+		hash: localIdentHash,
+		local,
+		contentHash,
+		chunkGraph,
+		module
+	});
+
+	// TODO move these things into interpolate
+	if (/\[local\]/.test(localIdent)) {
+		localIdent = localIdent.replace(/\[local\]/g, local);
+	}
+
+	if (/\[uniqueName\]/.test(localIdent)) {
+		localIdent = localIdent.replace(
+			/\[uniqueName\]/g,
+			/** @type {string} */ (uniqueName)
+		);
+	}
+
+	// Protect the first character from unsupported values
+	return localIdent.replace(/^((-?\d)|--)/, "_$1");
+};
+
+/**
+ * Computes the interpolated identifier for `(module, value, exportType)`.
+ * Module-level reference so `moduleGraph.cached` can use it as a stable
+ * computer key — repeated lookups during a build skip `cssExportConvention`,
+ * `getLocalIdent` (with its content / path hashing) and `escapeIdentifier`.
+ * @param {ModuleGraph} _moduleGraph module graph (unused, kept for `cached` signature)
+ * @param {CssModule} module css module the value resolves in
+ * @param {string} value raw value to interpolate
+ * @param {ExportType} exportType export type discriminator
+ * @param {ChunkGraph} chunkGraph chunk graph
+ * @param {RuntimeTemplate} runtimeTemplate runtime template
+ * @returns {string} interpolated identifier
+ */
+const computeInterpolatedIdentifier = (
+	_moduleGraph,
+	module,
+	value,
+	exportType,
+	chunkGraph,
+	runtimeTemplate
+) => {
+	const generator = /** @type {CssGenerator} */ (module.generator);
+	const local = cssExportConvention(
+		value,
+		/** @type {CssGeneratorExportsConvention} */
+		(generator.options.exportsConvention)
+	)[0];
+	const prefix = exportType === EXPORT_TYPE.CUSTOM_VARIABLE ? "--" : "";
+	return (
+		prefix +
+		getCssTokens().escapeIdentifier(
+			getLocalIdent(local, module, chunkGraph, runtimeTemplate),
+			runtimeTemplate.compilation.compiler.root
+		)
+	);
+};
+
+/**
+ * One `:export { name: value }` style entry — a plain object so a module holds
+ * a single `Dependency` plus a flat array instead of thousands of `Dependency`
+ * instances. `_conventionNames`/`_valueConventionNames` memoize the convention
+ * expansion per entry.
  * @typedef {object} CssExportEntry
  * @property {string} name export name
  * @property {Value} value value or [localName, importName, request?]
@@ -323,9 +492,10 @@ const findExportEntry = (module, name) => {
 };
 
 /**
- * Finds the active `CssIcssImportDependency` for a given local name. See the
- * legacy `CssIcssExportDependency.Template.findImportDep` for the `request`
- * disambiguation rules.
+ * Finds the active `CssIcssImportDependency` for a given local name. When a
+ * `request` is provided the lookup also requires the import dependency's
+ * `request` to match, so references between two `@value foo from "..."`
+ * declarations resolve through the import in scope at the reference site.
  * @param {Iterable<Dependency>} dependencies module dependencies to search
  * @param {string} localName local name
  * @param {string=} request user request of the `@value` import to match
@@ -502,7 +672,7 @@ CssExportsDependency.Template = class CssExportsDependencyTemplate extends (
 		const { moduleGraph, module, chunkGraph, runtimeTemplate } =
 			templateContext;
 		return moduleGraph.cached(
-			CssIcssExportDependency.computeInterpolatedIdentifier,
+			computeInterpolatedIdentifier,
 			/** @type {CssModule} */ (module),
 			value,
 			entry.exportType,
@@ -629,6 +799,12 @@ CssExportsDependency.Template = class CssExportsDependencyTemplate extends (
 		}
 	}
 };
+
+/** @type {Record<"NONE" | "REPLACE" | "APPEND" | "ONCE" | "SELF_REFERENCE", ExportMode>} */
+CssExportsDependency.EXPORT_MODE = EXPORT_MODE;
+
+/** @type {Record<"NORMAL" | "CUSTOM_VARIABLE" | "GRID_CUSTOM_IDENTIFIER" | "COMPOSES", ExportType>} */
+CssExportsDependency.EXPORT_TYPE = EXPORT_TYPE;
 
 makeSerializable(
 	CssExportsDependency,
