+++ lib/dependencies/CssIcssImportDependency.js
@@ -23,8 +23,8 @@ const CssImportDependency = require("./CssImportDependency");
 /** @typedef {import("../serialization/ObjectMiddleware").ObjectSerializerContext} ObjectSerializerContext */
 /** @typedef {import("../Dependency").ReferencedExports} ReferencedExports */
 /** @typedef {import("../../declarations/WebpackOptions").CssGeneratorExportsConvention} CssGeneratorExportsConvention */
-/** @typedef {import("./CssIcssExportDependency").ExportMode} ExportMode */
-/** @typedef {import("./CssIcssExportDependency").ExportType} ExportType */
+/** @typedef {import("./CssExportsDependency").ExportMode} ExportMode */
+/** @typedef {import("./CssExportsDependency").ExportType} ExportType */
 
 class CssIcssImportDependency extends CssImportDependency {
 	/**
