+++ lib/util/internalSerializables.js
@@ -76,8 +76,6 @@ module.exports = {
 		require("../dependencies/CssUrlDependency"),
 	"dependencies/CssIcssImportDependency": () =>
 		require("../dependencies/CssIcssImportDependency"),
-	"dependencies/CssIcssExportDependency": () =>
-		require("../dependencies/CssIcssExportDependency"),
 	"dependencies/CssExportsDependency": () =>
 		require("../dependencies/CssExportsDependency"),
 	"dependencies/CssIcssSymbolDependency": () =>
